+++ R/alignResource.R
@@ -5,13 +5,20 @@
 #' then rescales the search volumes of all fish species so that their resource
 #' encounter rate stays the same. This will therefore modify the encounter rate
 #' with other prey.
-#' 
+#'
+#' This is a generic function with a method for objects of class
+#' [MizerParams][mizer::MizerParams].
+#'
 #' @param params A MizerParams object
+#' @param ... Not used.
 #' @return A MizerParams object with updated resource carrying capacity,
 #'   resource initial value, resource parameter lambda and species parameter
 #'   gamma.
 #' @export
-alignResource <- function(params) {
+alignResource <- function(params, ...) UseMethod("alignResource")
+
+#' @export
+alignResource.MizerParams <- function(params, ...) {
     
     sc <- colSums(params@initial_n) * 
         params@w ^ params@resource_params$lambda