+++ R/plotBiomassVsSpecies.R
@@ -1,8 +1,19 @@
 #' Plot the biomass against species
-#' 
+#'
+#' This is a generic function with a method for objects of class
+#' [MizerParams][mizer::MizerParams].
+#'
 #' @param params A MizerParams object
+#' @param ... Not used.
+#' @examples
+#' plotBiomassVsSpecies(NS_params)
+#'
+
 #' @export
-plotBiomassVsSpecies <- function(params) {
+plotBiomassVsSpecies <- function(params, ...) UseMethod("plotBiomassVsSpecies")
+
+#' @export
+plotBiomassVsSpecies.MizerParams <- function(params, ...) {
     no_sp <- length(params@species_params$species)
     cutoff <- params@species_params$biomass_cutoff
     # When no cutoff known, set it to 0
@@ -12,7 +23,7 @@ plotBiomassVsSpecies <- function(params) {
     if (is.null(observed)) observed <- rep(NA, no_sp)
     
     # selector for foreground species
-    foreground <- !is.na(params@A)
+    foreground <- !params@species_params$is_background
     foreground_indices <- (1:no_sp)[foreground]
     biomass_model <- foreground_indices  # create vector of right length
     for (i in seq_along(foreground_indices)) {