+++ R/singleSpeciesSteady.R
@@ -10,7 +10,10 @@
 #' The result of applying this function is of course not a multi-species steady
 #' state, because after changing the abundances of the selected species the
 #' growth and death rates will have changed.
-#' 
+#'
+#' This is a generic function with a method for objects of class
+#' [MizerParams][mizer::MizerParams].
+#'
 #' @param params A MizerParams object
 #' @param species The species to be selected. Optional. By default all target
 #'   species are selected. A vector of species names, or a numeric vector with
@@ -20,12 +23,19 @@
 #'   choices are "egg" which keeps the egg density constant, "biomass" which 
 #'   keeps the total biomass of the species constant and "number" which keeps
 #'   the total number of individuals constant.
+#' @param ... Not used.
 #' @return A MizerParams object in which the initial abundances of the selected
 #'   species are changed to their single-species steady state abundances.
 #' @keywords internal
+
 #' @export
 singleSpeciesSteady <- function(params, species = NULL,
-                                keep = c("egg", "biomass", "number")) {
+                                keep = c("egg", "biomass", "number"), ...)
+    UseMethod("singleSpeciesSteady")
+
+#' @export
+singleSpeciesSteady.MizerParams <- function(params, species = NULL,
+                                keep = c("egg", "biomass", "number"), ...) {
     
     lifecycle::deprecate_warn("2.4.0", "singleSpeciesSteady()", 
                               "mizer::steadySingleSpecies()")